chore: remove dead code (LocationPicker)
