+++ packages/web-app-files/tests/unit/components/AppBar/CreateAndUpload.spec.js
@@ -310,6 +310,7 @@ function createStore(state = { currentFolder: {} }, fileHandlers = []) {
         },
         getters: {
           currentFolder: () => state.currentFolder,
+          clipboardResources: () => [],
           selectedFiles: () => []
         }
       }
