+++ tests/acceptance/features/webUIFilesCopy/copy.feature
@@ -127,40 +127,3 @@ Feature: copy files and folders
     And folder "simple-empty-folder" should be listed on the webUI
     And as "Alice" folder "folder with space/simple-empty-folder/simple-empty-folder" should exist in the server
     And as "Alice" folder "simple-empty-folder" should exist in the server
-
-  Scenario: cancel copying a file
-    Given user "Alice" has uploaded file "data.zip" to "data.zip" in the server
-    And user "Alice" has created folder "simple-empty-folder" in the server
-    And user "Alice" has logged in using the webUI
-    And the user has browsed to the personal page
-    When the user selects copy action for folder "data.zip" using the webUI
-    And the user selects the folder "simple-empty-folder" as a place to copy the file using the webUI
-    And the user cancels the attempt to copy resources using the webUI
-    Then file "data.zip" should be listed on the webUI
-    But  file "data.zip" should not be listed in the folder "simple-empty-folder" on the webUI
-
-  Scenario: cancel copying of multiple files at once
-    Given user "Alice" has uploaded file "data.zip" to "data.zip" in the server
-    And user "Alice" has uploaded file "new-data.zip" to "testapp.zip" in the server
-    And user "Alice" has created file "lorem.txt" in the server
-    And user "Alice" has created folder "simple-empty-folder" in the server
-    And user "Alice" has logged in using the webUI
-    And the user has browsed to the personal page
-    When the user marks these files for batch action using the webUI
-      | file_name   |
-      | data.zip    |
-      | lorem.txt   |
-      | testapp.zip |
-    And the user selects the copy button to copy files using the webUI
-    And the user selects the folder "simple-empty-folder" as a place to copy the files using the webUI
-    And the user cancels the attempt to copy resources using the webUI
-    Then the following files should be listed on the webUI
-      | file_name   |
-      | data.zip    |
-      | lorem.txt   |
-      | testapp.zip |
-    But these resources should not be listed in the folder "simple-empty-folder" on the webUI
-      | file_name   |
-      | data.zip    |
-      | lorem.txt   |
-      | testapp.zip |
